[fix] searx.httpclient: clean up
