+++ searx/httpclient/utils.py
@@ -61,6 +61,33 @@ def to_native_string(string, encoding='ascii'):
     return out
 
 
+# https://github.com/psf/requests/blob/0357b0724bd01cc89169fd40ca0132cf3248aff3/requests/utils.py#L287
+def to_key_val_list(value):
+    """Take an object and test to see if it can be represented as a
+    dictionary. If it can be, return a list of tuples, e.g.,
+    ::
+        >>> to_key_val_list([('key', 'val')])
+        [('key', 'val')]
+        >>> to_key_val_list({'key': 'val'})
+        [('key', 'val')]
+        >>> to_key_val_list('string')
+        Traceback (most recent call last):
+        ...
+        ValueError: cannot encode objects that are not 2-tuples
+    :rtype: list
+    """
+    if value is None:
+        return None
+
+    if isinstance(value, (str, bytes, bool, int)):
+        raise ValueError('cannot encode objects that are not 2-tuples')
+
+    if isinstance(value, Mapping):
+        value = value.items()
+
+    return list(value)
+
+
 # from https://github.com/psf/requests/blob/0357b0724bd01cc89169fd40ca0132cf3248aff3/requests/structures.py#L15
 class CaseInsensitiveDict(MutableMapping):
     """A case-insensitive ``dict``-like object.
